Extract code snippet into stand alone component
